rename resturant to restaurant and fix project configuration
