+++ app/views/restaurants/index.html.erb
@@ -1,4 +1,5 @@
 <p id="notice"><%= notice %></p>
+<%= javascript_pack_tag 'hello_react' %>
 
 <h1>Restaurants</h1>
