+++ config/environments/development.rb
@@ -54,4 +54,7 @@ config.webpacker.check_yarn_integrity = true
   # Use an evented file watcher to asynchronously detect changes in source code,
   # routes, locales, etc. This feature depends on the listen gem.
   config.file_watcher = ActiveSupport::EventedFileUpdateChecker
+
+  config.x.webpacker[:dev_server_host] = "http://127.0.0.1:8080"
+
 end
