+++ config/webpacker.yml
@@ -14,7 +14,6 @@ default: &default
   cache_manifest: false
 
   extensions:
-    - .jsx
     - .js
     - .sass
     - .scss
