+++ package.json
@@ -3,12 +3,26 @@
   "private": true,
   "dependencies": {
     "@rails/webpacker": "3.5",
-    "babel-preset-react": "^6.24.1",
+    "deep-extend": "^0.5.1",
     "prop-types": "^15.6.1",
     "react": "^16.3.2",
-    "react-dom": "^16.3.2"
+    "react-dom": "^16.3.2",
+    "redux": "^4.0.0"
   },
   "devDependencies": {
-    "webpack-dev-server": "2.11.2"
+    "babel": "^6.23.0",
+    "babel-core": "^6.26.3",
+    "babel-eslint": "^8.2.3",
+    "babel-loader": "^7.1.4",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-0": "^6.24.1",
+    "eslint": "^4.19.1",
+    "eslint-nibble": "^4.2.1",
+    "eslint-plugin-import": "^2.11.0",
+    "eslint-plugin-jsx-a11y": "^6.0.3",
+    "eslint-plugin-react": "^7.8.1",
+    "webpack": "^4.8.3",
+    "webpack-dev-server": "^3.1.4"
   }
 }
