+++ test/controllers/restaurants_controller_test.rb
@@ -1,0 +1,48 @@
+require 'test_helper'
+
+class RestaurantsControllerTest < ActionDispatch::IntegrationTest
+  setup do
+    @restaurant = restaurants(:one)
+  end
+
+  test "should get index" do
+    get restaurants_url
+    assert_response :success
+  end
+
+  test "should get new" do
+    get new_restaurant_url
+    assert_response :success
+  end
+
+  test "should create restaurant" do
+    assert_difference('Restaurant.count') do
+      post restaurants_url, params: { restaurant: { tenbis_enabled: @restaurant.tenbis_enabled, address: @restaurant.address, cuisine: @restaurant.cuisine, last_delivery_time: @restaurant.last_delivery_time, name: @restaurant.name } }
+    end
+
+    assert_redirected_to restaurant_url(Restaurant.last)
+  end
+
+  test "should show restaurant" do
+    get restaurant_url(@restaurant)
+    assert_response :success
+  end
+
+  test "should get edit" do
+    get edit_restaurant_url(@restaurant)
+    assert_response :success
+  end
+
+  test "should update restaurant" do
+    patch restaurant_url(@restaurant), params: { restaurant: { tenbis_enabled: @restaurant.tenbis_enabled, address: @restaurant.address, cuisine: @restaurant.cuisine, last_delivery_time: @restaurant.last_delivery_time, name: @restaurant.name } }
+    assert_redirected_to restaurant_url(@restaurant)
+  end
+
+  test "should destroy restaurant" do
+    assert_difference('Restaurant.count', -1) do
+      delete restaurant_url(@restaurant)
+    end
+
+    assert_redirected_to restaurants_url
+  end
+end
