+++ test/fixtures/reviews.yml
@@ -5,11 +5,11 @@ one:
   rating: 1
   comment: MyString
   deleted: false
-  resturant: one
+  restaurant: one
 
 two:
   name: MyString
   rating: 1
   comment: MyString
   deleted: false
-  resturant: two
+  restaurant: two
