+++ test/system/restaurants_test.rb
@@ -4,6 +4,6 @@ class RestaurantsTest < ApplicationSystemTestCase
   # test "visiting the index" do
   #   visit restaurants_url
   #
-  #   assert_selector "h1", text: "Resturant"
+  #   assert_selector "h1", text: "Restaurant"
   # end
 end
